Make infrastructure engine pluggable
Right now we have a hard binding to savanna.service.instances
Since we want to introduce a new Heat engine and keep the old one
for a while, we need a way to switch between them depending on
user's decision. The change makes it possible to replace
savanna.service.instances with another implementation by changing
a single config parameter.

Additionally that change updates savanna.conf.sample-full which
was not done for some time.

Implements blueprint infra-provisioning-extensions

Change-Id: I3408529948537bef8338a4927f17fa6d47ec2b76