Add script to report uncovered new lines
Let's update the coverage job behavior - it will run coverage on
current master and then it will run coverage on Head and compare
number of new lines w/o test coverage. The default number of
allowed uncovered lines is 4. If more lines uncovered job will
fail. We'll keep this job non-voting, but this script should help
us a lot to monitor the degradation of test coverage.

Change-Id: I16336cee3570f50bf80bc84ddb081737c1dd9dab