Remove log translations
Log messages are no longer being translated. This removes all use of
the _LE, _LI, _LW and _LC translation markers to simplify logging and to
avoid confusion with new contributions.

See:
http://lists.openstack.org/pipermail/openstack-i18n/2016-November/002574.html
http://lists.openstack.org/pipermail/openstack-dev/2017-March/113365.html

It will be a giant patch if all these _LE, _LI, _LW and _LC being deleted
within one patch, so this patch only delete _LW,  _LC and _LE has been
handled in previous patch; deletion of _LI will be in the following patches
related with this one.

Change-Id: Ia696ee76f22e488b9a219305f4f220eca2161cee